requirements.txt
Django==3.2.24
python-telegram-bot==20.4
mysqlclient==2.1.1 